Psec 1903 object lock feature

+++ test/s3_module_acceptance_test.go
@@ -1,12 +1,14 @@
 package test
 
 import (
+	"bytes"
 	"context"
+	"crypto/md5"
+	"encoding/base64"
 	"fmt"
 	"github.com/aws/aws-sdk-go-v2/feature/s3/manager"
 	"github.com/aws/aws-sdk-go-v2/service/s3/types"
 	test_structure "github.com/gruntwork-io/terratest/modules/test-structure"
-	"github.com/hmrc/terraform-aws-s3-bucket-standard/test/randomreader"
 	"log"
 	"strings"
 	"testing"
@@ -30,12 +32,14 @@ func TestReadRole(t *testing.T) {
 	ctx := context.Background()
 
 	testKey := "testS3KeyName"
+	body := "banana"
 	terraformOptions := copyTerraformAndReturnOptions(t, "examples/simple", map[string]interface{}{})
 	defer terraform.Destroy(t, terraformOptions)
 
 	terraform.InitAndApply(t, terraformOptions)
 
 	bucketName := terraform.Output(t, terraformOptions, "bucket_name")
+	objectLock := terraform.Output(t, terraformOptions, "object_lock")
 
 	versionId := CreateTestObject(t, ctx, bucketName, testKey)
 
@@ -50,17 +54,21 @@ func TestReadRole(t *testing.T) {
 	assert.Error(t, err)
 
 	_, err = readS3Client.PutObject(ctx, &s3.PutObjectInput{
-		Bucket: aws.String(bucketName),
-		Key:    aws.String(testKey),
-		Body:   strings.NewReader("banana")})
-	require.Error(t, err)
-
-	_, err = readS3Client.DeleteObject(ctx, &s3.DeleteObjectInput{
-		VersionId: versionId,
-		Bucket:    aws.String(bucketName),
-		Key:       &testKey,
+		Bucket:     aws.String(bucketName),
+		Key:        aws.String(testKey),
+		Body:       strings.NewReader(body),
+		ContentMD5: aws.String(md5EncodeBody(body)),
 	})
 	require.Error(t, err)
+
+	if objectLock != "true" {
+		_, err = readS3Client.DeleteObject(ctx, &s3.DeleteObjectInput{
+			VersionId: versionId,
+			Bucket:    aws.String(bucketName),
+			Key:       &testKey,
+		})
+		require.Error(t, err)
+	}
 
 	_, err = readS3Client.ListBucketInventoryConfigurations(ctx, &s3.ListBucketInventoryConfigurationsInput{
 		Bucket: aws.String(bucketName),
@@ -78,19 +86,23 @@ func TestWriteRole(t *testing.T) {
 	ctx := context.Background()
 
 	testKey := "testS3KeyName"
+	body := "banana"
 	terraformOptions := copyTerraformAndReturnOptions(t, "examples/simple", map[string]interface{}{})
 	defer terraform.Destroy(t, terraformOptions)
 
 	terraform.InitAndApply(t, terraformOptions)
 
 	bucketName := terraform.Output(t, terraformOptions, "bucket_name")
+	objectLock := terraform.Output(t, terraformOptions, "object_lock")
 
 	//test write role
 	writeS3Client := S3ClientFromOutputArn(t, ctx, terraformOptions, "write_role_arn")
 	uploadResponse, err := writeS3Client.PutObject(ctx, &s3.PutObjectInput{
-		Bucket: aws.String(bucketName),
-		Key:    aws.String(testKey),
-		Body:   strings.NewReader("banana")})
+		Bucket:     aws.String(bucketName),
+		Key:        aws.String(testKey),
+		Body:       strings.NewReader(body),
+		ContentMD5: aws.String(md5EncodeBody(body)),
+	})
 	require.NoError(t, err)
 
 	_, err = writeS3Client.GetObject(ctx, &s3.GetObjectInput{
@@ -113,13 +125,15 @@ func TestWriteRole(t *testing.T) {
 	assert.Error(t, err)
 
 	//test delete on write role
-	_, err = writeS3Client.DeleteObject(ctx, &s3.DeleteObjectInput{
-		VersionId: uploadResponse.VersionId,
-		Bucket:    aws.String(bucketName),
-		Key:       &testKey,
-	})
-	fmt.Println(err)
-	require.NoError(t, err)
+	if objectLock != "true" {
+		_, err = writeS3Client.DeleteObject(ctx, &s3.DeleteObjectInput{
+			VersionId: uploadResponse.VersionId,
+			Bucket:    aws.String(bucketName),
+			Key:       &testKey,
+		})
+		fmt.Println(err)
+		require.NoError(t, err)
+	}
 }
 
 func TestListRole(t *testing.T) {
@@ -127,15 +141,20 @@ func TestListRole(t *testing.T) {
 	ctx := context.Background()
 
 	testKey := "testS3KeyName"
+	body := "banana"
 	terraformOptions := copyTerraformAndReturnOptions(t, "examples/simple", map[string]interface{}{})
 	defer terraform.Destroy(t, terraformOptions)
 
 	terraform.InitAndApply(t, terraformOptions)
 
 	bucketName := terraform.Output(t, terraformOptions, "bucket_name")
+	objectLock := terraform.Output(t, terraformOptions, "object_lock")
 
 	versionId := CreateTestObject(t, ctx, bucketName, testKey)
-	defer DeleteTestObject(t, ctx, bucketName, testKey, versionId)
+
+	if objectLock != "true" {
+		defer DeleteTestObject(t, ctx, bucketName, testKey, versionId)
+	}
 
 	listS3Client := S3ClientFromOutputArn(t, ctx, terraformOptions, "list_role_arn")
 	ListResponse, err := listS3Client.ListObjectsV2(ctx, &s3.ListObjectsV2Input{Bucket: aws.String(bucketName)})
@@ -150,9 +169,11 @@ func TestListRole(t *testing.T) {
 	assert.Error(t, err)
 
 	_, err = listS3Client.PutObject(ctx, &s3.PutObjectInput{
-		Bucket: aws.String(bucketName),
-		Key:    aws.String(testKey),
-		Body:   strings.NewReader("banana")})
+		Bucket:     aws.String(bucketName),
+		Key:        aws.String(testKey),
+		Body:       strings.NewReader(body),
+		ContentMD5: aws.String(md5EncodeBody(body)),
+	})
 	require.Error(t, err)
 
 	_, err = listS3Client.GetObject(ctx, &s3.GetObjectInput{
@@ -172,15 +193,19 @@ func TestAdminRole(t *testing.T) {
 	ctx := context.Background()
 
 	testKey := "testS3KeyName"
+	body := "banana"
 	terraformOptions := copyTerraformAndReturnOptions(t, "examples/simple", map[string]interface{}{})
 	defer terraform.Destroy(t, terraformOptions)
 
 	terraform.InitAndApply(t, terraformOptions)
 
 	bucketName := terraform.Output(t, terraformOptions, "bucket_name")
+	objectLock := terraform.Output(t, terraformOptions, "object_lock")
 
 	versionId := CreateTestObject(t, ctx, bucketName, testKey)
-	defer DeleteTestObject(t, ctx, bucketName, testKey, versionId)
+	if objectLock != "true" {
+		defer DeleteTestObject(t, ctx, bucketName, testKey, versionId)
+	}
 
 	adminS3Client := S3ClientFromOutputArn(t, ctx, terraformOptions, "admin_role_arn")
 	_, err := adminS3Client.ListBucketInventoryConfigurations(ctx, &s3.ListBucketInventoryConfigurationsInput{
@@ -203,17 +228,21 @@ func TestAdminRole(t *testing.T) {
 	assert.Error(t, err)
 
 	_, err = adminS3Client.PutObject(ctx, &s3.PutObjectInput{
-		Bucket: aws.String(bucketName),
-		Key:    aws.String(testKey),
-		Body:   strings.NewReader("banana")})
-	require.Error(t, err)
-
-	_, err = adminS3Client.DeleteObject(ctx, &s3.DeleteObjectInput{
-		VersionId: versionId,
-		Bucket:    aws.String(bucketName),
-		Key:       &testKey,
+		Bucket:     aws.String(bucketName),
+		Key:        aws.String(testKey),
+		Body:       strings.NewReader(body),
+		ContentMD5: aws.String(md5EncodeBody(body)),
 	})
 	require.Error(t, err)
+
+	if objectLock != "true" {
+		_, err = adminS3Client.DeleteObject(ctx, &s3.DeleteObjectInput{
+			VersionId: versionId,
+			Bucket:    aws.String(bucketName),
+			Key:       &testKey,
+		})
+		require.Error(t, err)
+	}
 }
 
 func TestMetadataRole(t *testing.T) {
@@ -221,16 +250,19 @@ func TestMetadataRole(t *testing.T) {
 	ctx := context.Background()
 
 	testKey := "testS3KeyName"
+	body := "banana"
 	terraformOptions := copyTerraformAndReturnOptions(t, "examples/simple", map[string]interface{}{})
 	defer terraform.Destroy(t, terraformOptions)
 
 	terraform.InitAndApply(t, terraformOptions)
 
 	bucketName := terraform.Output(t, terraformOptions, "bucket_name")
+	objectLock := terraform.Output(t, terraformOptions, "object_lock")
 
 	versionId := CreateTestObject(t, ctx, bucketName, testKey)
-	defer DeleteTestObject(t, ctx, bucketName, testKey, versionId)
-
+	if objectLock != "true" {
+		defer DeleteTestObject(t, ctx, bucketName, testKey, versionId)
+	}
 	metadataS3Client := S3ClientFromOutputArn(t, ctx, terraformOptions, "metadata_role_arn")
 	_, err := metadataS3Client.GetBucketAccelerateConfiguration(ctx, &s3.GetBucketAccelerateConfigurationInput{
 		Bucket: aws.String(bucketName),
@@ -247,23 +279,28 @@ func TestMetadataRole(t *testing.T) {
 	assert.Error(t, err)
 
 	_, err = metadataS3Client.PutObject(ctx, &s3.PutObjectInput{
-		Bucket: aws.String(bucketName),
-		Key:    aws.String(testKey),
-		Body:   strings.NewReader("banana")})
-	require.Error(t, err)
-
-	_, err = metadataS3Client.DeleteObject(ctx, &s3.DeleteObjectInput{
-		VersionId: versionId,
-		Bucket:    aws.String(bucketName),
-		Key:       &testKey,
+		Bucket:     aws.String(bucketName),
+		Key:        aws.String(testKey),
+		Body:       strings.NewReader(body),
+		ContentMD5: aws.String(md5EncodeBody(body)),
 	})
 	require.Error(t, err)
+
+	if objectLock != "true" {
+		_, err = metadataS3Client.DeleteObject(ctx, &s3.DeleteObjectInput{
+			VersionId: versionId,
+			Bucket:    aws.String(bucketName),
+			Key:       &testKey,
+		})
+		require.Error(t, err)
+	}
 }
 
 func TestUploadMultiPart(t *testing.T) {
 	t.Parallel()
 	ctx := context.Background()
 
+	body := "large_multi_part_upload"
 	testKey := "testS3KeyName"
 	terraformOptions := copyTerraformAndReturnOptions(t, "examples/simple", map[string]interface{}{})
 	defer terraform.Destroy(t, terraformOptions)
@@ -277,9 +314,10 @@ func TestUploadMultiPart(t *testing.T) {
 		u.PartSize = manager.MinUploadPartSize
 	})
 	_, err := uploader.Upload(ctx, &s3.PutObjectInput{
-		Bucket: &bucketName,
-		Key:    &testKey,
-		Body:   randomreader.New(manager.MinUploadPartSize + 100),
+		Bucket:     &bucketName,
+		Key:        &testKey,
+		Body:       bytes.NewReader([]byte(body)),
+		ContentMD5: aws.String(md5EncodeBody(body)),
 	})
 	require.NoError(t, err)
 }
@@ -296,46 +334,58 @@ func TestCannotUseDifferentKeys(t *testing.T) {
 	bucketKeyId := terraform.Output(t, terraformOptions, "bucket_kms_key_id")
 	additionalKey := terraform.Output(t, terraformOptions, "additional_kms_key_arn")
 
+	var body string
 	cfg := CreateConfig(t, ctx)
 	client := s3.NewFromConfig(cfg)
+	body = "different kms key"
 	_, err := client.PutObject(ctx, &s3.PutObjectInput{
 		Bucket:               &bucketName,
 		Key:                  aws.String("differentKey"),
-		Body:                 strings.NewReader("different kms key"),
+		Body:                 strings.NewReader(body),
+		ContentMD5:           aws.String(md5EncodeBody(body)),
 		ServerSideEncryption: types.ServerSideEncryptionAwsKms,
 		SSEKMSKeyId:          &additionalKey})
 	require.Error(t, err)
+	body = "service key"
 	_, err = client.PutObject(ctx, &s3.PutObjectInput{
 		Bucket:               &bucketName,
 		Key:                  aws.String("serviceKey"),
-		Body:                 strings.NewReader("service key"),
+		Body:                 strings.NewReader(body),
+		ContentMD5:           aws.String(md5EncodeBody(body)),
 		ServerSideEncryption: types.ServerSideEncryptionAwsKms})
 	require.Error(t, err)
+	body = "AES"
 	_, err = client.PutObject(ctx, &s3.PutObjectInput{
 		Bucket:               &bucketName,
 		Key:                  aws.String("aesKey"),
-		Body:                 strings.NewReader("AES"),
+		Body:                 strings.NewReader(body),
+		ContentMD5:           aws.String(md5EncodeBody(body)),
 		ServerSideEncryption: types.ServerSideEncryptionAes256})
 	require.Error(t, err)
-
+	body = "Dont specify any"
 	putOut, err := client.PutObject(ctx, &s3.PutObjectInput{
-		Bucket: &bucketName,
-		Key:    aws.String("noEncSpec"),
-		Body:   strings.NewReader("Dont specify any")})
+		Bucket:     &bucketName,
+		Key:        aws.String("noEncSpec"),
+		Body:       strings.NewReader(body),
+		ContentMD5: aws.String(md5EncodeBody(body))})
 	require.NoError(t, err)
 	assert.Equal(t, bucketKeyArn, *putOut.SSEKMSKeyId)
+	body = "specify alg and bucket key"
 	putOut, err = client.PutObject(ctx, &s3.PutObjectInput{
 		Bucket:               &bucketName,
 		Key:                  aws.String("specKeyByArn"),
-		Body:                 strings.NewReader("specify alg and bucket key"),
+		Body:                 strings.NewReader(body),
+		ContentMD5:           aws.String(md5EncodeBody(body)),
 		ServerSideEncryption: types.ServerSideEncryptionAwsKms,
 		SSEKMSKeyId:          &bucketKeyArn})
 	require.NoError(t, err)
 	assert.Equal(t, bucketKeyArn, *putOut.SSEKMSKeyId)
+	body = "specify alg and bucket key"
 	putOut, err = client.PutObject(ctx, &s3.PutObjectInput{
 		Bucket:               &bucketName,
 		Key:                  aws.String("specKeyById"),
-		Body:                 strings.NewReader("specify alg and bucket key"),
+		Body:                 strings.NewReader(body),
+		ContentMD5:           aws.String(md5EncodeBody(body)),
 		ServerSideEncryption: types.ServerSideEncryptionAwsKms,
 		SSEKMSKeyId:          &bucketKeyId})
 	require.NoError(t, err)
@@ -383,10 +433,12 @@ func CreateConfig(t *testing.T, ctx context.Context) aws.Config {
 
 func CreateTestObject(t *testing.T, ctx context.Context, bucketName string, testKey string) *string {
 	terraformClient := s3.NewFromConfig(CreateConfig(t, ctx))
+	body := "banana"
 	uploadResponse, Err := terraformClient.PutObject(ctx, &s3.PutObjectInput{
-		Bucket: aws.String(bucketName),
-		Key:    aws.String(testKey),
-		Body:   strings.NewReader("banana")})
+		Bucket:     aws.String(bucketName),
+		Key:        aws.String(testKey),
+		Body:       strings.NewReader(body),
+		ContentMD5: aws.String(md5EncodeBody(body))})
 	require.NoError(t, Err)
 	return uploadResponse.VersionId
 }
@@ -420,4 +472,11 @@ func CopyTerraformAndReturnOptions(t *testing.T, pathFromRootToSource string, va
 		TerraformDir: tempTestFolder,
 		Vars:         vars,
 	})
+}
+
+func md5EncodeBody(s string) string {
+	hash := md5.Sum([]byte(s))
+	var data []byte
+	data = hash[:]
+	return base64.StdEncoding.EncodeToString(data)
 }
